DSBC: PASAMOS a long la Key

+++ src/main/java/com/conmigo/app/bbdd/dao/UserDao.java
@@ -2,7 +2,7 @@ package com.conmigo.app.bbdd.dao;
 
 import com.conmigo.app.bbdd.jpa.User;
 
-public interface UserDao extends GenericDao<User, Integer> {
+public interface UserDao extends GenericDao<User, Long> {
 
 	User findByName( String username );
